Address the case of horizontal line intersection.
The bounding box for the resulting line had a lower-left and top-right
height of the same value (zero difference). This made intersection points
with any rounding error in the Y axis unable to fall inside the box.
Permit some uncertainty in the Y value in these cases to allow for
rounding error. Do the same for the X axis (vertical lines).

Signed-off-by: Tinkerer <tinkerer@zappem.net>