Improve inside counting further
Consider four points: A and (B,C,D). If A.Y = C.Y then
a line of fixed Y component to the right of A.X is going
to intersect with BC and CD, at C. The inside calculation
was counting this as two intersections which tricked the
algorithm into not recognizing this as an opportunity for
an odd number of crossings. Whether or not C should be
counted once or twice in such an overlap calculation is
if B.Y and D.Y are on opposite sides of A.Y: if they are
we should count this as a single intercept; if not, then
we should count this as two intercepts.

Signed-off-by: Tinkerer <tinkerer@zappem.net>