Improve Inside calculation for polygon and handle -0.0
The Point.Inside(*Shape) was confused when there were lines
in the shape that were of the same Y value as the Point. Fix
this by skipping all such points.

Also, we make use of a map[Point] to note points of intersection
Go has inconsistent handling of -0.0 vs 0.0 in such a map[] with
the == comparison. Namely (-0.0,1.0) and (0.0,1.0) index different
map entries, but == compare as the same. This was confusing some
of the outline recognizer code, so we no longer use == for Point
comparison, but use Point.NotSame(Point) instead.

Signed-off-by: Tinkerer <tinkerer@zappem.net>