Fixed Inside test, and rounding error handling
The prior update to Point.Inside(Shape) missed an important corner case.
Added an example of this situation to TestInside, and resolved that.

Further, improve the consistency of handling near-zero situations.
Before, we had the Zeroish value which we were using to both compare
linear nearness to zero, and also quadratic nearness to zero. This led
to some problems combining polygons that had a lot of colinearity. Added
a new test TestRound and used it to debug when to use Zeroish and when
to use Zeroish^2 for rounding error tolerant comparisons.

Signed-off-by: Tinkerer <tinkerer@zappem.net>