Resolved the concave Union problem: TestConcaveE()
Introduce (*Shape).Hull() and (*Shape).Duplicate() methods.

When considering if two polygons are overlapping, with one wholly
inside the other, reject that as correct if the points of overlap are
non-hull points of the otherwise swallowed up polygon.

The `rowmant` font `t` problem was an upstream issue with the `hershey`
font utf8 mappings. https://github.com/tinkerator/hershey/issues/1 .
I also suspect that my `J` concern in this font is also part of the same
issue, so I'm including an observation about that in the same bug report.

Signed-off-by: Tinkerer <tinkerer@zappem.net>