Improve intersect() and clean up hole handling
This removed the spurious dots visible in T and Z (and z) in some
hershey font renderings to polygons. I'm still seeing a bit of
rendering corruption. So, not quite ready for 1.0.0 yet.

Adjusted one test that was aggressively testing for the prior rounding
error compensation. This compensation was partially the cause of the
problems addressed with this commit.

Signed-off-by: Tinkerer <tinkerer@zappem.net>