Exact bounding box matching less fragile now
When calculating if a line intersection was within the range
of a pair of line segments, the intersect() function was too
strict when determining if a bounding box had zero width or
height. Relax this comparison with comparison to Zeroish.

This took care of some glyph corruption that proved to be
sensitive to location in coordinate space. Add a TestSticks
test for this, which was derived from Hershey font rendering of a
`#`.

Signed-off-by: Tinkerer <tinkerer@zappem.net>